Refactor logging format helpers
